backend/pyproject.toml
[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Backend for the Todo App"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlmodel",
    "asyncpg",
    "aiosqlite",
    "alembic",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]",
    "pydantic-settings",
    "slowapi",
    "python-json-logger",
    "uvicorn[standard]",
    "psycopg2-binary",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.20",
    "python-dateutil>=2.9.0.post0",
    "apscheduler>=3.11.1",
    "aiosmtplib>=5.0.0",
    "pywebpush>=2.1.2",
    "sentry-sdk>=2.47.0",
    "redis>=7.1.0",
    "aiocache>=0.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "pip-audit",
    "pre-commit",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104", "D107"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.8",
]
